optimize: database queries with findByIdAndUserId method
- Add findByIdAndUserId method to PlaidItemRepository interface and implementation
- Create institution.service.ts with getInstitutionForUser and parseInstitutionId methods
- Update institution controller to use optimized service methods
- Remove duplicate validation and ownership verification logic
- Single query now handles both ID lookup and ownership verification

Performance: Reduces database calls from 2 to 1 for institution verification
Maintainability: Centralizes validation logic in reusable service methods
